+++ changelog.md
@@ -1,3 +1,8 @@
+#### 0.4.31 - 7/20/2019 ####
+Extreme Dazzle casino upgrade<br>
+Metallurgist minor trait<br>
+Gambler minor trait<br>
+
 #### 0.4.30 - 7/19/2019 ####
 Portuguese translation by Rodrigodd<br>
 DNA Sequencer upgrade<br>